+++ Sources/TextMarkupKit/MiniMarkdownGrammar.swift
@@ -66,7 +66,7 @@ public final class MiniMarkdownGrammar: PackratGrammar {
   }
 
   /// Singleton for convenience.
-  public static let shared = MiniMarkdownGrammar()
+  @MainActor public static let shared = MiniMarkdownGrammar()
 
   public private(set) lazy var start: ParsingRule = block.memoize()
     .repeating(0...)