+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - `ParsedString.parsedContents` for seeing exactly how the parse tree looks for a ParsedString.
 
+### Changed
+
+- `SyntaxTreeNode.path(to:)` now works when the `location == endIndex`. This used to be an invalid parameter. Now, it is valid and will get associated with the last child in the tree.
+
 ### Fixed 
 
 - Fixed bug where maintaining the `AttributesArray` could result in negative-length runs